feat: add license const
