+++ server/controller/common/const.go
@@ -260,6 +260,8 @@ const (
 	AGENT_LICENSE_FUNCTION_ASSET_K8S
 	AGENT_LICENSE_FUNCTION_ASSET_CMDB
 	AGENT_LICENSE_FUNCTION_AI_COPILOT
+	AGENT_LICENSE_FUNCTION_BIZ_SCENE
+	AGENT_LICENSE_FUNCTION_BIZ_SYSTEM
 	AGENT_LICENSE_FUNCTION_MAX
 )
