+++ server/controller/tagrecorder/ch_custom_biz_service_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/deepflowio/deepflow/server/controller/config"
+	"github.com/deepflowio/deepflow/server/controller/config/common"
 	"github.com/deepflowio/deepflow/server/controller/db/metadb"
 	metadbmodel "github.com/deepflowio/deepflow/server/controller/db/metadb/model"
 )
@@ -136,7 +137,7 @@ func TestChCustomBizService_generateNewData(t *testing.T) {
 
 			// Setup config
 			cfg := config.ControllerConfig{
-				DFWebService: config.DFWebService{
+				DFWebService: common.DFWebService{
 					Enabled: tt.dfWebServiceEnabled,
 					Host:    host,
 					Port:    port,
